requirements.txt
redis>=3.2,<3.3
tornado>=6.0,<6.1
slackeventsapi>=2.1,<2.2
python-dotenv>=0.10,<0.11
slackclient>=2.0,<2.1
